+++ Shared/src/main/java/arch/galaxyeclipse/shared/common/GePosition.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class GePosition {
-    private float x;
-    private float y;
+    private int x;
+    private int y;
 }
